KAFKA-14675: Extract metadata-related tasks from Fetcher into MetadataFetcher
