+++ clients/src/test/java/org/apache/kafka/clients/consumer/KafkaConsumerTest.java
@@ -30,6 +30,7 @@ import org.apache.kafka.clients.consumer.internals.ConsumerNetworkClient;
 import org.apache.kafka.clients.consumer.internals.ConsumerProtocol;
 import org.apache.kafka.clients.consumer.internals.Fetcher;
 import org.apache.kafka.clients.consumer.internals.MockRebalanceListener;
+import org.apache.kafka.clients.consumer.internals.MetadataFetcher;
 import org.apache.kafka.clients.consumer.internals.SubscriptionState;
 import org.apache.kafka.common.Cluster;
 import org.apache.kafka.common.IsolationLevel;
@@ -2626,6 +2627,8 @@ public class KafkaConsumerTest {
                 throwOnStableOffsetNotSupported,
                 null);
         }
+        IsolationLevel isolationLevel = IsolationLevel.READ_UNCOMMITTED;
+        ApiVersions apiVersions = new ApiVersions();
         Fetcher<String, String> fetcher = new Fetcher<>(
                 loggerFactory,
                 consumerClient,
@@ -2643,10 +2646,17 @@ public class KafkaConsumerTest {
                 metrics,
                 metricsRegistry.fetcherMetrics,
                 time,
+                isolationLevel,
+                apiVersions);
+        MetadataFetcher metadataFetcher = new MetadataFetcher(loggerFactory,
+                consumerClient,
+                metadata,
+                subscription,
+                time,
                 retryBackoffMs,
                 requestTimeoutMs,
-                IsolationLevel.READ_UNCOMMITTED,
-                new ApiVersions());
+                isolationLevel,
+                apiVersions);
 
         return new KafkaConsumer<>(
                 loggerFactory,
@@ -2655,6 +2665,7 @@ public class KafkaConsumerTest {
                 keyDeserializer,
                 deserializer,
                 fetcher,
+                metadataFetcher,
                 interceptors,
                 time,
                 consumerClient,
