+++ clients/src/test/java/org/apache/kafka/clients/consumer/internals/FetcherTest.java
@@ -204,6 +204,7 @@ public class FetcherTest {
     private ApiVersions apiVersions = new ApiVersions();
     private ConsumerNetworkClient consumerClient;
     private Fetcher<?, ?> fetcher;
+    private MetadataFetcher metadataFetcher;
 
     private MemoryRecords records;
     private MemoryRecords nextRecords;
@@ -1919,7 +1920,7 @@ public class FetcherTest {
         assignFromUser(singleton(tp0));
         subscriptions.seek(tp0, 5L);
 
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
         assertFalse(client.hasInFlightRequests());
         assertTrue(subscriptions.isFetchable(tp0));
         assertEquals(5, subscriptions.position(tp0).offset);
@@ -1933,7 +1934,7 @@ public class FetcherTest {
 
         client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.EARLIEST_TIMESTAMP,
             validLeaderEpoch), listOffsetResponse(Errors.NONE, 1L, 5L));
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
         consumerClient.pollNoWakeup();
         assertFalse(subscriptions.isOffsetResetNeeded(tp0));
         assertTrue(subscriptions.isFetchable(tp0));
@@ -1950,7 +1951,7 @@ public class FetcherTest {
 
         client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.LATEST_TIMESTAMP),
                 listOffsetResponse(Errors.NONE, 1L, 5L));
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
         consumerClient.pollNoWakeup();
         assertFalse(subscriptions.isOffsetResetNeeded(tp0));
         assertTrue(subscriptions.isFetchable(tp0));
@@ -1969,7 +1970,7 @@ public class FetcherTest {
         // Fail with OFFSET_NOT_AVAILABLE
         client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.LATEST_TIMESTAMP,
             validLeaderEpoch), listOffsetResponse(Errors.OFFSET_NOT_AVAILABLE, 1L, 5L), false);
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
         consumerClient.pollNoWakeup();
         assertFalse(subscriptions.hasValidPosition(tp0));
         assertTrue(subscriptions.isOffsetResetNeeded(tp0));
@@ -1979,7 +1980,7 @@ public class FetcherTest {
         time.sleep(retryBackoffMs);
         client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.LATEST_TIMESTAMP,
             validLeaderEpoch), listOffsetResponse(Errors.LEADER_NOT_AVAILABLE, 1L, 5L), false);
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
         consumerClient.pollNoWakeup();
         assertFalse(subscriptions.hasValidPosition(tp0));
         assertTrue(subscriptions.isOffsetResetNeeded(tp0));
@@ -1989,7 +1990,7 @@ public class FetcherTest {
         time.sleep(retryBackoffMs);
         client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.LATEST_TIMESTAMP),
                 listOffsetResponse(Errors.NONE, 1L, 5L), false);
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
         consumerClient.pollNoWakeup();
         assertTrue(subscriptions.hasValidPosition(tp0));
         assertFalse(subscriptions.isOffsetResetNeeded(tp0));
@@ -2018,7 +2019,7 @@ public class FetcherTest {
             ListOffsetsRequest request = (ListOffsetsRequest) body;
             return request.isolationLevel() == isolationLevel;
         }, listOffsetResponse(Errors.NONE, 1L, 5L));
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
         consumerClient.pollNoWakeup();
 
         assertFalse(subscriptions.isOffsetResetNeeded(tp0));
@@ -2036,7 +2037,7 @@ public class FetcherTest {
         client.updateMetadata(initialUpdateResponse);
         Node node = initialUpdateResponse.brokers().iterator().next();
         client.backoff(node, 500);
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
         assertEquals(0, consumerClient.pendingRequestCount());
         consumerClient.pollNoWakeup();
         assertTrue(subscriptions.isOffsetResetNeeded(tp0));
@@ -2045,7 +2046,7 @@ public class FetcherTest {
         time.sleep(500);
         client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.EARLIEST_TIMESTAMP),
                 listOffsetResponse(Errors.NONE, 1L, 5L));
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
         consumerClient.pollNoWakeup();
 
         assertFalse(subscriptions.isOffsetResetNeeded(tp0));
@@ -2061,7 +2062,7 @@ public class FetcherTest {
 
         client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.EARLIEST_TIMESTAMP,
             validLeaderEpoch), listOffsetResponse(Errors.NONE, 1L, 5L));
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
         consumerClient.pollNoWakeup();
 
         assertFalse(subscriptions.isOffsetResetNeeded(tp0));
@@ -2078,7 +2079,7 @@ public class FetcherTest {
         // First fetch fails with stale metadata
         client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.LATEST_TIMESTAMP,
             validLeaderEpoch), listOffsetResponse(Errors.NOT_LEADER_OR_FOLLOWER, 1L, 5L), false);
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
         consumerClient.pollNoWakeup();
         assertFalse(subscriptions.hasValidPosition(tp0));
 
@@ -2091,7 +2092,7 @@ public class FetcherTest {
         time.sleep(retryBackoffMs);
         client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.LATEST_TIMESTAMP),
                 listOffsetResponse(Errors.NONE, 1L, 5L));
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
         consumerClient.pollNoWakeup();
 
         assertFalse(subscriptions.isOffsetResetNeeded(tp0));
@@ -2113,7 +2114,7 @@ public class FetcherTest {
         subscriptions.requestOffsetReset(tp0, OffsetResetStrategy.LATEST);
         client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.LATEST_TIMESTAMP),
                 listOffsetResponse(tp0, Errors.NONE, 1L, 5L, 1));
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
         consumerClient.pollNoWakeup();
 
         // Reset should be satisfied and no metadata update requested
@@ -2138,7 +2139,7 @@ public class FetcherTest {
         // Now we see a ListOffsets with leaderEpoch=2 epoch, we trigger a metadata update
         client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.LATEST_TIMESTAMP, 1),
                 listOffsetResponse(tp0, Errors.NONE, 1L, 5L, 2));
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
         consumerClient.pollNoWakeup();
 
         assertFalse(subscriptions.isOffsetResetNeeded(tp0));
@@ -2155,7 +2156,7 @@ public class FetcherTest {
         // First request gets a disconnect
         client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.LATEST_TIMESTAMP,
             validLeaderEpoch), listOffsetResponse(Errors.NONE, 1L, 5L), true);
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
         consumerClient.pollNoWakeup();
         assertFalse(subscriptions.hasValidPosition(tp0));
 
@@ -2165,7 +2166,7 @@ public class FetcherTest {
         assertFalse(client.hasPendingMetadataUpdates());
 
         // No retry until the backoff passes
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
         consumerClient.pollNoWakeup();
         assertFalse(client.hasInFlightRequests());
         assertFalse(subscriptions.hasValidPosition(tp0));
@@ -2174,7 +2175,7 @@ public class FetcherTest {
         time.sleep(retryBackoffMs);
         client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.LATEST_TIMESTAMP),
                 listOffsetResponse(Errors.NONE, 1L, 5L));
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
         consumerClient.pollNoWakeup();
 
         assertFalse(subscriptions.isOffsetResetNeeded(tp0));
@@ -2189,7 +2190,7 @@ public class FetcherTest {
         subscriptions.requestOffsetReset(tp0, OffsetResetStrategy.LATEST);
 
         // Send the ListOffsets request to reset the position
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
         consumerClient.pollNoWakeup();
         assertFalse(subscriptions.hasValidPosition(tp0));
         assertTrue(client.hasInFlightRequests());
@@ -2213,7 +2214,7 @@ public class FetcherTest {
         subscriptions.requestOffsetReset(tp0, OffsetResetStrategy.LATEST);
 
         // Send the ListOffsets request to reset the position
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
         consumerClient.pollNoWakeup();
         assertFalse(subscriptions.hasValidPosition(tp0));
         assertTrue(client.hasInFlightRequests());
@@ -2260,7 +2261,7 @@ public class FetcherTest {
         assignFromUser(singleton(tp0));
 
         subscriptions.requestOffsetReset(tp0, OffsetResetStrategy.EARLIEST);
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
 
         client.prepareResponse(req -> {
             if (listOffsetMatchesExpectedReset(tp0, OffsetResetStrategy.EARLIEST, req)) {
@@ -2277,7 +2278,7 @@ public class FetcherTest {
         assertTrue(subscriptions.isOffsetResetNeeded(tp0));
         assertEquals(OffsetResetStrategy.LATEST, subscriptions.resetStrategy(tp0));
 
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
         client.prepareResponse(
             req -> listOffsetMatchesExpectedReset(tp0, OffsetResetStrategy.LATEST, req),
             listOffsetResponse(Errors.NONE, 1L, 10L)
@@ -2295,7 +2296,7 @@ public class FetcherTest {
         subscriptions.requestOffsetReset(tp0, OffsetResetStrategy.LATEST);
 
         // Send the ListOffsets request to reset the position
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
         consumerClient.pollNoWakeup();
         assertFalse(subscriptions.hasValidPosition(tp0));
         assertTrue(client.hasInFlightRequests());
@@ -2320,7 +2321,7 @@ public class FetcherTest {
         subscriptions.requestOffsetReset(tp0, OffsetResetStrategy.LATEST);
 
         // Send the ListOffsets request to reset the position
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
         consumerClient.pollNoWakeup();
         assertFalse(subscriptions.hasValidPosition(tp0));
         assertTrue(client.hasInFlightRequests());
@@ -2345,19 +2346,19 @@ public class FetcherTest {
         // First request gets a disconnect
         client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.LATEST_TIMESTAMP,
             validLeaderEpoch), listOffsetResponse(Errors.TOPIC_AUTHORIZATION_FAILED, -1, -1), false);
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
         consumerClient.pollNoWakeup();
         assertFalse(subscriptions.hasValidPosition(tp0));
 
         try {
-            fetcher.resetOffsetsIfNeeded();
+            metadataFetcher.resetOffsetsIfNeeded();
             fail("Expected authorization error to be raised");
         } catch (TopicAuthorizationException e) {
             assertEquals(singleton(tp0.topic()), e.unauthorizedTopics());
         }
 
         // The exception should clear after being raised, but no retry until the backoff
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
         consumerClient.pollNoWakeup();
         assertFalse(client.hasInFlightRequests());
         assertFalse(subscriptions.hasValidPosition(tp0));
@@ -2366,7 +2367,7 @@ public class FetcherTest {
         time.sleep(retryBackoffMs);
         client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.LATEST_TIMESTAMP),
                 listOffsetResponse(Errors.NONE, 1L, 5L));
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
         consumerClient.pollNoWakeup();
 
         assertFalse(subscriptions.isOffsetResetNeeded(tp0));
@@ -2384,7 +2385,7 @@ public class FetcherTest {
         assertTrue(subscriptions.isFetchable(tp0));
 
         subscriptions.markPendingRevocation(singleton(tp0));
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
 
         // once a partition is marked pending, it should not be fetchable
         assertFalse(subscriptions.isOffsetResetNeeded(tp0));
@@ -2412,7 +2413,7 @@ public class FetcherTest {
 
         client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.LATEST_TIMESTAMP,
             validLeaderEpoch), listOffsetResponse(Errors.NONE, 1L, 10L));
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
         consumerClient.pollNoWakeup();
 
         assertFalse(subscriptions.isOffsetResetNeeded(tp0));
@@ -2428,7 +2429,7 @@ public class FetcherTest {
         subscriptions.requestOffsetReset(tp0);
         subscriptions.pause(tp0); // paused partition does not have a valid position
 
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
         consumerClient.pollNoWakeup();
 
         assertTrue(subscriptions.isOffsetResetNeeded(tp0));
@@ -2443,7 +2444,7 @@ public class FetcherTest {
         subscriptions.seek(tp0, 10);
         subscriptions.pause(tp0); // paused partition already has a valid position
 
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
 
         assertFalse(subscriptions.isOffsetResetNeeded(tp0));
         assertFalse(subscriptions.isFetchable(tp0)); // because tp is paused
@@ -2458,7 +2459,7 @@ public class FetcherTest {
         assignFromUser(singleton(tp0));
         client.prepareResponse(newMetadataResponse(topicName, Errors.NONE));
 
-        Map<String, List<PartitionInfo>> allTopics = fetcher.getAllTopicMetadata(time.timer(5000L));
+        Map<String, List<PartitionInfo>> allTopics = metadataFetcher.getAllTopicMetadata(time.timer(5000L));
 
         assertEquals(initialUpdateResponse.topicMetadata().size(), allTopics.size());
     }
@@ -2470,7 +2471,7 @@ public class FetcherTest {
         assignFromUser(singleton(tp0));
         client.prepareResponse(null, true);
         client.prepareResponse(newMetadataResponse(topicName, Errors.NONE));
-        Map<String, List<PartitionInfo>> allTopics = fetcher.getAllTopicMetadata(time.timer(5000L));
+        Map<String, List<PartitionInfo>> allTopics = metadataFetcher.getAllTopicMetadata(time.timer(5000L));
         assertEquals(initialUpdateResponse.topicMetadata().size(), allTopics.size());
     }
 
@@ -2479,7 +2480,7 @@ public class FetcherTest {
         // since no response is prepared, the request should timeout
         buildFetcher();
         assignFromUser(singleton(tp0));
-        assertThrows(TimeoutException.class, () -> fetcher.getAllTopicMetadata(time.timer(50L)));
+        assertThrows(TimeoutException.class, () -> metadataFetcher.getAllTopicMetadata(time.timer(50L)));
     }
 
     @Test
@@ -2488,7 +2489,7 @@ public class FetcherTest {
         assignFromUser(singleton(tp0));
         client.prepareResponse(newMetadataResponse(topicName, Errors.TOPIC_AUTHORIZATION_FAILED));
         try {
-            fetcher.getAllTopicMetadata(time.timer(10L));
+            metadataFetcher.getAllTopicMetadata(time.timer(10L));
             fail();
         } catch (TopicAuthorizationException e) {
             assertEquals(singleton(topicName), e.unauthorizedTopics());
@@ -2500,7 +2501,7 @@ public class FetcherTest {
         buildFetcher();
         assignFromUser(singleton(tp0));
         client.prepareResponse(newMetadataResponse(topicName, Errors.INVALID_TOPIC_EXCEPTION));
-        assertThrows(InvalidTopicException.class, () -> fetcher.getTopicMetadata(
+        assertThrows(InvalidTopicException.class, () -> metadataFetcher.getTopicMetadata(
                 new MetadataRequest.Builder(Collections.singletonList(topicName), true), time.timer(5000L)));
     }
 
@@ -2510,7 +2511,7 @@ public class FetcherTest {
         assignFromUser(singleton(tp0));
         client.prepareResponse(newMetadataResponse(topicName, Errors.UNKNOWN_TOPIC_OR_PARTITION));
 
-        Map<String, List<PartitionInfo>> topicMetadata = fetcher.getTopicMetadata(
+        Map<String, List<PartitionInfo>> topicMetadata = metadataFetcher.getTopicMetadata(
                 new MetadataRequest.Builder(Collections.singletonList(topicName), true), time.timer(5000L));
         assertNull(topicMetadata.get(topicName));
     }
@@ -2522,7 +2523,7 @@ public class FetcherTest {
         client.prepareResponse(newMetadataResponse(topicName, Errors.LEADER_NOT_AVAILABLE));
         client.prepareResponse(newMetadataResponse(topicName, Errors.NONE));
 
-        Map<String, List<PartitionInfo>> topicMetadata = fetcher.getTopicMetadata(
+        Map<String, List<PartitionInfo>> topicMetadata = metadataFetcher.getTopicMetadata(
                 new MetadataRequest.Builder(Collections.singletonList(topicName), true), time.timer(5000L));
         assertTrue(topicMetadata.containsKey(topicName));
     }
@@ -2567,7 +2568,7 @@ public class FetcherTest {
         client.prepareResponse(altered);
 
         Map<String, List<PartitionInfo>> topicMetadata =
-            fetcher.getTopicMetadata(new MetadataRequest.Builder(Collections.singletonList(topicName), false),
+            metadataFetcher.getTopicMetadata(new MetadataRequest.Builder(Collections.singletonList(topicName), false),
                     time.timer(5000L));
 
         assertNotNull(topicMetadata);
@@ -2981,7 +2982,7 @@ public class FetcherTest {
     @Test
     public void testGetOffsetsForTimesTimeout() {
         buildFetcher();
-        assertThrows(TimeoutException.class, () -> fetcher.offsetsForTimes(
+        assertThrows(TimeoutException.class, () -> metadataFetcher.offsetsForTimes(
             Collections.singletonMap(new TopicPartition(topicName, 2), 1000L), time.timer(100L)));
     }
 
@@ -2990,7 +2991,7 @@ public class FetcherTest {
         buildFetcher();
 
         // Empty map
-        assertTrue(fetcher.offsetsForTimes(new HashMap<>(), time.timer(100L)).isEmpty());
+        assertTrue(metadataFetcher.offsetsForTimes(new HashMap<>(), time.timer(100L)).isEmpty());
         // Unknown Offset
         testGetOffsetsForTimesWithUnknownOffset();
         // Error code none with unknown offset
@@ -3017,7 +3018,7 @@ public class FetcherTest {
         subscriptions.requestOffsetReset(tp0, OffsetResetStrategy.LATEST);
 
         client.prepareResponse(listOffsetResponse(Errors.FENCED_LEADER_EPOCH, 1L, 5L));
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
         consumerClient.pollNoWakeup();
 
         assertTrue(subscriptions.isOffsetResetNeeded(tp0));
@@ -3125,7 +3126,7 @@ public class FetcherTest {
             }, listOffsetResponse(tp1, Errors.NONE, fetchTimestamp, 5L), newLeader);
 
             Map<TopicPartition, OffsetAndTimestamp> offsetAndTimestampMap =
-                fetcher.offsetsForTimes(
+                metadataFetcher.offsetsForTimes(
                     Utils.mkMap(Utils.mkEntry(tp0, fetchTimestamp),
                     Utils.mkEntry(tp1, fetchTimestamp)), time.timer(Integer.MAX_VALUE));
 
@@ -3148,7 +3149,7 @@ public class FetcherTest {
         subscriptions.requestOffsetReset(tp0, OffsetResetStrategy.LATEST);
 
         client.prepareResponse(listOffsetResponse(Errors.UNKNOWN_LEADER_EPOCH, 1L, 5L));
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
         consumerClient.pollNoWakeup();
 
         assertTrue(subscriptions.isOffsetResetNeeded(tp0));
@@ -3171,7 +3172,7 @@ public class FetcherTest {
 
         // Request latest offset
         subscriptions.requestOffsetReset(tp0);
-        fetcher.resetOffsetsIfNeeded();
+        metadataFetcher.resetOffsetsIfNeeded();
 
         // Check for epoch in outgoing request
         MockClient.RequestMatcher matcher = body -> {
@@ -3227,11 +3228,11 @@ public class FetcherTest {
         timestampToSearch.put(tp1, ListOffsetsRequest.LATEST_TIMESTAMP);
         timestampToSearch.put(t2p0, ListOffsetsRequest.LATEST_TIMESTAMP);
         Map<TopicPartition, OffsetAndTimestamp> offsetAndTimestampMap =
-            fetcher.offsetsForTimes(timestampToSearch, time.timer(Long.MAX_VALUE));
+            metadataFetcher.offsetsForTimes(timestampToSearch, time.timer(Long.MAX_VALUE));
 
-        assertNotNull(offsetAndTimestampMap.get(tp0), "Expect Fetcher.offsetsForTimes() to return non-null result for " + tp0);
-        assertNotNull(offsetAndTimestampMap.get(tp1), "Expect Fetcher.offsetsForTimes() to return non-null result for " + tp1);
-        assertNotNull(offsetAndTimestampMap.get(t2p0), "Expect Fetcher.offsetsForTimes() to return non-null result for " + t2p0);
+        assertNotNull(offsetAndTimestampMap.get(tp0), "Expect MetadataFetcher.offsetsForTimes() to return non-null result for " + tp0);
+        assertNotNull(offsetAndTimestampMap.get(tp1), "Expect MetadataFetcher.offsetsForTimes() to return non-null result for " + tp1);
+        assertNotNull(offsetAndTimestampMap.get(t2p0), "Expect MetadataFetcher.offsetsForTimes() to return non-null result for " + t2p0);
         assertEquals(11L, offsetAndTimestampMap.get(tp0).offset());
         assertEquals(32L, offsetAndTimestampMap.get(tp1).offset());
         assertEquals(54L, offsetAndTimestampMap.get(t2p0).offset());
@@ -3262,9 +3263,9 @@ public class FetcherTest {
 
         Map<TopicPartition, Long> timestampToSearch = new HashMap<>();
         timestampToSearch.put(tp0, ListOffsetsRequest.LATEST_TIMESTAMP);
-        Map<TopicPartition, OffsetAndTimestamp> offsetAndTimestampMap = fetcher.offsetsForTimes(timestampToSearch, time.timer(Long.MAX_VALUE));
+        Map<TopicPartition, OffsetAndTimestamp> offsetAndTimestampMap = metadataFetcher.offsetsForTimes(timestampToSearch, time.timer(Long.MAX_VALUE));
 
-        assertNotNull(offsetAndTimestampMap.get(tp0), "Expect Fetcher.offsetsForTimes() to return non-null result for " + tp0);
+        assertNotNull(offsetAndTimestampMap.get(tp0), "Expect MetadataFetcher.offsetsForTimes() to return non-null result for " + tp0);
         assertEquals(11L, offsetAndTimestampMap.get(tp0).offset());
         assertNotNull(metadata.fetch().partitionCountForTopic(anotherTopic));
     }
@@ -3281,7 +3282,7 @@ public class FetcherTest {
         offsetsToExpect.put(tp0, null);
         offsetsToExpect.put(tp1, null);
 
-        assertEquals(offsetsToExpect, fetcher.offsetsForTimes(offsetsToSearch, time.timer(0)));
+        assertEquals(offsetsToExpect, metadataFetcher.offsetsForTimes(offsetsToSearch, time.timer(0)));
     }
 
     @Test
@@ -3309,7 +3310,7 @@ public class FetcherTest {
         offsetsToSearch.put(tp0, ListOffsetsRequest.EARLIEST_TIMESTAMP);
         offsetsToSearch.put(tp1, ListOffsetsRequest.EARLIEST_TIMESTAMP);
 
-        assertThrows(TimeoutException.class, () -> fetcher.offsetsForTimes(offsetsToSearch, time.timer(1)));
+        assertThrows(TimeoutException.class, () -> metadataFetcher.offsetsForTimes(offsetsToSearch, time.timer(1)));
     }
 
     @Test
@@ -3847,6 +3848,7 @@ public class FetcherTest {
         LogContext logContext = new LogContext();
         buildDependencies(new MetricConfig(), Long.MAX_VALUE, new SubscriptionState(logContext, OffsetResetStrategy.EARLIEST), logContext);
 
+        IsolationLevel isolationLevel = IsolationLevel.READ_UNCOMMITTED;
         fetcher = new Fetcher<byte[], byte[]>(
                 new LogContext(),
                 consumerClient,
@@ -3864,9 +3866,7 @@ public class FetcherTest {
                 metrics,
                 metricsRegistry,
                 time,
-                retryBackoffMs,
-                requestTimeoutMs,
-                IsolationLevel.READ_UNCOMMITTED,
+                isolationLevel,
                 apiVersions) {
             @Override
             protected FetchSessionHandler sessionHandler(int id) {
@@ -3913,6 +3913,15 @@ public class FetcherTest {
                 }
             }
         };
+        this.metadataFetcher = new MetadataFetcher(logContext,
+                                                   consumerClient,
+                                                   metadata,
+                                                   subscriptions,
+                                                   time,
+                                                   retryBackoffMs,
+                                                   requestTimeoutMs,
+                                                   isolationLevel,
+                                                   apiVersions);
 
         MetadataResponse initialMetadataResponse = RequestTestUtils.metadataUpdateWithIds(1,
                 singletonMap(topicName, numPartitions), tp -> validLeaderEpoch, topicIds);
@@ -4151,7 +4160,7 @@ public class FetcherTest {
         timestampToSearch.put(t2p0, 0L);
         timestampToSearch.put(tp1, 0L);
         Map<TopicPartition, OffsetAndTimestamp> offsetAndTimestampMap =
-                fetcher.offsetsForTimes(timestampToSearch, time.timer(Long.MAX_VALUE));
+                metadataFetcher.offsetsForTimes(timestampToSearch, time.timer(Long.MAX_VALUE));
 
         if (expectedOffsetForP0 == null)
             assertNull(offsetAndTimestampMap.get(t2p0));
@@ -4190,7 +4199,7 @@ public class FetcherTest {
         Map<TopicPartition, Long> timestampToSearch = new HashMap<>();
         timestampToSearch.put(tp0, 0L);
         Map<TopicPartition, OffsetAndTimestamp> offsetAndTimestampMap =
-                fetcher.offsetsForTimes(timestampToSearch, time.timer(Long.MAX_VALUE));
+                metadataFetcher.offsetsForTimes(timestampToSearch, time.timer(Long.MAX_VALUE));
 
         assertTrue(offsetAndTimestampMap.containsKey(tp0));
         assertNull(offsetAndTimestampMap.get(tp0));
@@ -4200,7 +4209,7 @@ public class FetcherTest {
     public void testGetOffsetsForTimesWithUnknownOffsetV0() {
         buildFetcher();
         // Empty map
-        assertTrue(fetcher.offsetsForTimes(new HashMap<>(), time.timer(100L)).isEmpty());
+        assertTrue(metadataFetcher.offsetsForTimes(new HashMap<>(), time.timer(100L)).isEmpty());
         // Unknown Offset
         client.reset();
         // Ensure metadata has both partition.
@@ -4227,7 +4236,7 @@ public class FetcherTest {
         Map<TopicPartition, Long> timestampToSearch = new HashMap<>();
         timestampToSearch.put(tp0, 0L);
         Map<TopicPartition, OffsetAndTimestamp> offsetAndTimestampMap =
-                fetcher.offsetsForTimes(timestampToSearch, time.timer(Long.MAX_VALUE));
+                metadataFetcher.offsetsForTimes(timestampToSearch, time.timer(Long.MAX_VALUE));
 
         assertTrue(offsetAndTimestampMap.containsKey(tp0));
         assertNull(offsetAndTimestampMap.get(tp0));
@@ -4334,7 +4343,7 @@ public class FetcherTest {
 
         assertEquals(subscriptions.assignedPartitions(), allRequestedPartitions);
 
-        fetcher.validateOffsetsIfNeeded();
+        metadataFetcher.validateOffsetsIfNeeded();
         consumerClient.pollNoWakeup();
 
         assertTrue(subscriptions.assignedPartitions()
@@ -4365,7 +4374,7 @@ public class FetcherTest {
         assertTrue(subscriptions.awaitingValidation(tp0));
 
         // No version information is initially available, but the node is now connected
-        fetcher.validateOffsetsIfNeeded();
+        metadataFetcher.validateOffsetsIfNeeded();
         assertTrue(subscriptions.awaitingValidation(tp0));
         assertTrue(client.isConnected(node.idString()));
         apiVersions.update(node.idString(), NodeApiVersions.create());
@@ -4375,7 +4384,7 @@ public class FetcherTest {
             prepareOffsetsForLeaderEpochResponse(tp0, Errors.NONE, epochOne, 30L),
             node);
 
-        fetcher.validateOffsetsIfNeeded();
+        metadataFetcher.validateOffsetsIfNeeded();
         consumerClient.pollNoWakeup();
 
         assertFalse(subscriptions.awaitingValidation(tp0));
@@ -4414,7 +4423,7 @@ public class FetcherTest {
             // Update metadata to epoch=2, enter validation
             metadata.updateWithCurrentRequestVersion(RequestTestUtils.metadataUpdateWithIds("dummy", 1,
                     Collections.emptyMap(), partitionCounts, tp -> epochTwo, topicIds), false, 0L);
-            fetcher.validateOffsetsIfNeeded();
+            metadataFetcher.validateOffsetsIfNeeded();
 
             // Offset validation is skipped
             assertFalse(subscriptions.awaitingValidation(tp0));
@@ -4465,7 +4474,7 @@ public class FetcherTest {
         final short responseVersion = 8;
         metadata.updateWithCurrentRequestVersion(RequestTestUtils.metadataUpdateWith("dummy", 1,
             Collections.emptyMap(), partitionCounts, tp -> null, MetadataResponse.PartitionMetadata::new, responseVersion, topicIds), false, 0L);
-        fetcher.validateOffsetsIfNeeded();
+        metadataFetcher.validateOffsetsIfNeeded();
         // Offset validation is skipped
         assertFalse(subscriptions.awaitingValidation(tp0));
     }
@@ -4522,7 +4531,7 @@ public class FetcherTest {
         Metadata.LeaderAndEpoch leaderAndEpoch = new Metadata.LeaderAndEpoch(metadata.currentLeader(tp0).leader, Optional.of(epochOne));
         subscriptions.seekUnvalidated(tp0, new SubscriptionState.FetchPosition(initialOffset, Optional.of(epochOne), leaderAndEpoch));
 
-        fetcher.validateOffsetsIfNeeded();
+        metadataFetcher.validateOffsetsIfNeeded();
 
         consumerClient.poll(time.timer(Duration.ZERO));
         assertTrue(subscriptions.awaitingValidation(tp0));
@@ -4535,7 +4544,7 @@ public class FetcherTest {
 
         if (offsetResetStrategy == OffsetResetStrategy.NONE) {
             LogTruncationException thrown =
-                assertThrows(LogTruncationException.class, () -> fetcher.validateOffsetsIfNeeded());
+                assertThrows(LogTruncationException.class, () -> metadataFetcher.validateOffsetsIfNeeded());
             assertEquals(singletonMap(tp0, initialOffset), thrown.offsetOutOfRangePartitions());
 
             if (endOffset == UNDEFINED_EPOCH_OFFSET || leaderEpoch == UNDEFINED_EPOCH) {
@@ -4547,7 +4556,7 @@ public class FetcherTest {
             }
             assertTrue(subscriptions.awaitingValidation(tp0));
         } else {
-            fetcher.validateOffsetsIfNeeded();
+            metadataFetcher.validateOffsetsIfNeeded();
             assertFalse(subscriptions.awaitingValidation(tp0));
         }
     }
@@ -4572,7 +4581,7 @@ public class FetcherTest {
         Metadata.LeaderAndEpoch leaderAndEpoch = new Metadata.LeaderAndEpoch(metadata.currentLeader(tp0).leader, Optional.of(epochOne));
         subscriptions.seekUnvalidated(tp0, new SubscriptionState.FetchPosition(0, Optional.of(epochOne), leaderAndEpoch));
 
-        fetcher.validateOffsetsIfNeeded();
+        metadataFetcher.validateOffsetsIfNeeded();
         consumerClient.poll(time.timer(Duration.ZERO));
         assertTrue(subscriptions.awaitingValidation(tp0));
         assertTrue(client.hasInFlightRequests());
@@ -4617,7 +4626,7 @@ public class FetcherTest {
         // Update metadata to epoch=2, enter validation
         metadata.updateWithCurrentRequestVersion(RequestTestUtils.metadataUpdateWithIds("dummy", 1,
                 Collections.emptyMap(), partitionCounts, tp -> epochTwo, topicIds), false, 0L);
-        fetcher.validateOffsetsIfNeeded();
+        metadataFetcher.validateOffsetsIfNeeded();
         assertTrue(subscriptions.awaitingValidation(tp0));
 
         // Update the position to epoch=3, as we would from a fetch
@@ -4635,7 +4644,7 @@ public class FetcherTest {
         assertTrue(subscriptions.awaitingValidation(tp0), "Expected validation to fail since leader epoch changed");
 
         // Next round of validation, should succeed in validating the position
-        fetcher.validateOffsetsIfNeeded();
+        metadataFetcher.validateOffsetsIfNeeded();
         client.prepareResponse(prepareOffsetsForLeaderEpochResponse(tp0, Errors.NONE, epochThree, 10L));
         consumerClient.pollNoWakeup();
         assertFalse(subscriptions.awaitingValidation(tp0), "Expected validation to succeed with latest epoch");
@@ -4665,8 +4674,8 @@ public class FetcherTest {
         assertEquals(0, fetcher.sendFetches());
 
         // Complete reset and now we can fetch
-        fetcher.resetOffsetIfNeeded(tp0, OffsetResetStrategy.LATEST,
-                new Fetcher.ListOffsetData(100, 1L, Optional.empty()));
+        metadataFetcher.resetOffsetIfNeeded(tp0, OffsetResetStrategy.LATEST,
+                new MetadataFetcher.ListOffsetData(100, 1L, Optional.empty()));
         assertEquals(1, fetcher.sendFetches());
     }
 
@@ -4710,7 +4719,7 @@ public class FetcherTest {
         subscriptions.seekValidated(tp0, new SubscriptionState.FetchPosition(0, Optional.of(1), leaderAndEpoch));
 
         // Check for truncation, this should cause tp0 to go into validation
-        fetcher.validateOffsetsIfNeeded();
+        metadataFetcher.validateOffsetsIfNeeded();
 
         // No fetches sent since we entered validation
         assertEquals(0, fetcher.sendFetches());
@@ -4979,7 +4988,7 @@ public class FetcherTest {
         buildFetcher();
         assignFromUser(singleton(tp0));
         client.prepareResponse(listOffsetResponse(tp0, Errors.NONE, ListOffsetsRequest.EARLIEST_TIMESTAMP, 2L));
-        assertEquals(singletonMap(tp0, 2L), fetcher.beginningOffsets(singleton(tp0), time.timer(5000L)));
+        assertEquals(singletonMap(tp0, 2L), metadataFetcher.beginningOffsets(singleton(tp0), time.timer(5000L)));
     }
 
     @Test
@@ -4987,7 +4996,7 @@ public class FetcherTest {
         buildFetcher();
         assignFromUser(singleton(tp0));
         client.prepareResponse(listOffsetResponse(tp0, Errors.NONE, ListOffsetsRequest.EARLIEST_TIMESTAMP, 2L));
-        assertEquals(singletonMap(tp0, 2L), fetcher.beginningOffsets(asList(tp0, tp0), time.timer(5000L)));
+        assertEquals(singletonMap(tp0, 2L), metadataFetcher.beginningOffsets(asList(tp0, tp0), time.timer(5000L)));
     }
 
     @Test
@@ -4999,13 +5008,13 @@ public class FetcherTest {
         expectedOffsets.put(tp2, 6L);
         assignFromUser(expectedOffsets.keySet());
         client.prepareResponse(listOffsetResponse(expectedOffsets, Errors.NONE, ListOffsetsRequest.EARLIEST_TIMESTAMP, ListOffsetsResponse.UNKNOWN_EPOCH));
-        assertEquals(expectedOffsets, fetcher.beginningOffsets(asList(tp0, tp1, tp2), time.timer(5000L)));
+        assertEquals(expectedOffsets, metadataFetcher.beginningOffsets(asList(tp0, tp1, tp2), time.timer(5000L)));
     }
 
     @Test
     public void testBeginningOffsetsEmpty() {
         buildFetcher();
-        assertEquals(emptyMap(), fetcher.beginningOffsets(emptyList(), time.timer(5000L)));
+        assertEquals(emptyMap(), metadataFetcher.beginningOffsets(emptyList(), time.timer(5000L)));
     }
 
     @Test
@@ -5013,7 +5022,7 @@ public class FetcherTest {
         buildFetcher();
         assignFromUser(singleton(tp0));
         client.prepareResponse(listOffsetResponse(tp0, Errors.NONE, ListOffsetsRequest.LATEST_TIMESTAMP, 5L));
-        assertEquals(singletonMap(tp0, 5L), fetcher.endOffsets(singleton(tp0), time.timer(5000L)));
+        assertEquals(singletonMap(tp0, 5L), metadataFetcher.endOffsets(singleton(tp0), time.timer(5000L)));
     }
 
     @Test
@@ -5021,7 +5030,7 @@ public class FetcherTest {
         buildFetcher();
         assignFromUser(singleton(tp0));
         client.prepareResponse(listOffsetResponse(tp0, Errors.NONE, ListOffsetsRequest.LATEST_TIMESTAMP, 5L));
-        assertEquals(singletonMap(tp0, 5L), fetcher.endOffsets(asList(tp0, tp0), time.timer(5000L)));
+        assertEquals(singletonMap(tp0, 5L), metadataFetcher.endOffsets(asList(tp0, tp0), time.timer(5000L)));
     }
 
     @Test
@@ -5033,13 +5042,13 @@ public class FetcherTest {
         expectedOffsets.put(tp2, 9L);
         assignFromUser(expectedOffsets.keySet());
         client.prepareResponse(listOffsetResponse(expectedOffsets, Errors.NONE, ListOffsetsRequest.LATEST_TIMESTAMP, ListOffsetsResponse.UNKNOWN_EPOCH));
-        assertEquals(expectedOffsets, fetcher.endOffsets(asList(tp0, tp1, tp2), time.timer(5000L)));
+        assertEquals(expectedOffsets, metadataFetcher.endOffsets(asList(tp0, tp1, tp2), time.timer(5000L)));
     }
 
     @Test
     public void testEndOffsetsEmpty() {
         buildFetcher();
-        assertEquals(emptyMap(), fetcher.endOffsets(emptyList(), time.timer(5000L)));
+        assertEquals(emptyMap(), metadataFetcher.endOffsets(emptyList(), time.timer(5000L)));
     }
 
     private MockClient.RequestMatcher offsetsForLeaderEpochRequestMatcher(
@@ -5337,6 +5346,13 @@ public class FetcherTest {
                 subscriptions,
                 metrics,
                 metricsRegistry,
+                time,
+                isolationLevel,
+                apiVersions);
+        metadataFetcher = new MetadataFetcher(logContext,
+                consumerClient,
+                metadata,
+                subscriptions,
                 time,
                 retryBackoffMs,
                 requestTimeoutMs,
